requirements.txt
telethon
setuptools
cryptg
pillow
python-dotenv